Migrated to a cleaner authentication implementation
