+++ src/actions/index.js
@@ -4,13 +4,6 @@ import * as notebooks from "./notebooks";
 import * as selectedNav from "./selectedNav";
 import * as totalNotebookLength from "./totalNotebookLength";
 
-// Essentially an instance variable
-export let app = "";
-// Updates app
-export const updateApp = (newApp) => {
-    app = newApp;
-};
-
 export { authentication };
 export { getNotebooks };
 export { notebooks };
